Abstracted axios code
