+++ frontend/src/screens/home/Home.jsx
@@ -7,15 +7,7 @@ import { OnCampusIndicator } from "../../components/Dashboard/OnCampusIndicator"
 import { UserClaimView } from "../../components/User/UserClaimView";
 import { useNavigate } from "react-router-dom";
 import { useUser } from "../../contexts/userContext";
-import axios from "axios";
-
-axios.defaults.xsrfCookieName = 'csrftoken';
-axios.defaults.xsrfHeaderName = 'X-CSRFToken';
-axios.defaults.withCredentials = true;
-
-const client = axios.create({
-    baseURL: "http://127.0.0.1:8000"
-})
+import { client } from "../../../axios";
 
     // Based on state passed in which contains if logged in or not, display the home page with a login and register button
